+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.2.0]
+### Added
+- Root of Nightmares loot table.
+- Root of Nightmares craftables.
+
 ## [7.1.0]
 ### Added
 - New `/vendor ada-1` command, find your missing shaders.